devrequirements.txt
# Tooling
flake8
mypy
twine
sphinx
sphinx-autobuild
lxml-stubs